feat: impl lossless outfit rw
